content/057/services/signature/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aio-pika==9.4.3
clickhouse-connect==0.7.19
pydantic==2.8.2
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-aio-pika==0.47b0
pytest==8.3.2
pytest-asyncio==0.23.8
